+++ src/components/position.js
@@ -1,14 +1,25 @@
-
-
-/**
- * 
- * 	Component holding entity's position, width, and height.
- *  By convention, "position" is the bottom center of the entity's AABB
- * 
- */
-
+var vec3 = require('gl-vec3')
 
 export default function (noa) {
+
+    /**
+     * 
+     * 	Component holding entity's position, width, and height.
+     *  By convention, entity's "position" is the bottom center of its AABB
+     * 
+     *  Of the various properties, _localPosition is the "real", 
+     *  single-source-of-truth position. Others are derived.
+     *  Local coords are relative to `noa.worldOriginOffset`.
+     * 
+     *  Props:
+     *      position: pos in global coords (may be low precision)
+     *      _localPosition: precise pos in local coords
+     *      _renderPosition: [x,y,z] in LOCAL COORDS
+     *      _extents: array [lo, lo, lo, hi, hi, hi] in LOCAL COORDS
+     * 
+     */
+
+
 
     return {
 
@@ -17,27 +28,29 @@ export default function (noa) {
         order: 60,
 
         state: {
-            localPosition: null,
-            renderPosition: null,
-            width: +0,
-            height: +0,
+            position: null,
+            width: +1,
+            height: +1,
+            _localPosition: null,
+            _renderPosition: null,
             _extents: null,
-            _extentsChanged: true,
         },
 
 
         onAdd: function (eid, state) {
-            if (state.localPosition) {
-                var p = noa.vec3.create()
-                noa.vec3.copy(p, state.localPosition)
-                state.localPosition = p
-            } else {
-                state.localPosition = noa.vec3.create()
-            }
+            // copy position into a plain array
+            var pos = [0, 0, 0]
+            if (state.position) vec3.copy(pos, state.position)
+            state.position = pos
 
-            state.renderPosition = noa.vec3.create()
-            noa.vec3.copy(state.renderPosition, state.localPosition)
+            state._localPosition = vec3.create()
+            state._renderPosition = vec3.create()
             state._extents = new Float32Array(6)
+
+            // on init only, set local from global
+            noa.globalToLocal(state.position, null, state._localPosition)
+            vec3.copy(state._renderPosition, state._localPosition)
+            updatePositionExtents(state)
         },
 
         onRemove: null,
@@ -45,10 +58,10 @@ export default function (noa) {
 
 
         system: function (dt, states) {
+            var off = noa.worldOriginOffset
             states.forEach(state => {
-                if (!state._extentsChanged) return
-                updateExtents(state._extents, state.localPosition, state.height, state.width)
-                state._extentsChanged = false
+                vec3.add(state.position, state._localPosition, off)
+                updatePositionExtents(state)
             })
         },
 
@@ -57,12 +70,16 @@ export default function (noa) {
 }
 
 
-function updateExtents(ext, pos, height, width) {
-    var hw = width / 2
-    ext[0] = pos[0] - hw
-    ext[1] = pos[1]
-    ext[2] = pos[2] - hw
-    ext[3] = pos[0] + hw
-    ext[4] = pos[1] + height
-    ext[5] = pos[2] + hw
+
+// update an entity's position state `_extents` 
+export function updatePositionExtents(state) {
+    var hw = state.width / 2
+    var lpos = state._localPosition
+    var ext = state._extents
+    ext[0] = lpos[0] - hw
+    ext[1] = lpos[1]
+    ext[2] = lpos[2] - hw
+    ext[3] = lpos[0] + hw
+    ext[4] = lpos[1] + state.height
+    ext[5] = lpos[2] + hw
 }